# docs/openapi.yml
# THIS FILE IS GENERATED BY OpenApiSchemaExportTest — DO NOT EDIT MANUALLY.
# Run `./mvnw -pl service -am test -Dtest=OpenApiSchemaExportTest -Dsurefire.failIfNoSpecifiedTests=false` to regenerate.
---
openapi: 3.1.0
components:
  schemas:
    CreateEventRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          pattern: \S
        occurredAt:
          $ref: "#/components/schemas/Instant"
        payload:
          type: string
        description:
          type: string
        parentId:
          type: string
    EventDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        occurredAt:
          $ref: "#/components/schemas/Instant"
        payload:
          type: string
        description:
          type: string
        parentId:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    PublishEventRequest:
      type: object
      required:
      - name
      - occurredAt
      properties:
        name:
          type: string
          pattern: \S
        occurredAt:
          $ref: "#/components/schemas/Instant"
        payload:
          type: string
        description:
          type: string
        parentId:
          type: string
    Response:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/EventDto"
    Response1:
      type: object
      properties:
        names:
          type: array
          items:
            type: string
    Response2:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventDto"
        nextCursor:
          type: string
    Response3:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/EventDto"
    Response4:
      type: object
      properties:
        success:
          type: boolean
    Response5:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/EventDto"
paths:
  /events/api/events:
    get:
      parameters:
      - name: attr
        in: query
        schema:
          type: array
          items:
            type: string
      - name: cursor
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: string
      - name: name
        in: query
        schema:
          type: string
      - name: order
        in: query
        schema:
          type: string
      - name: parentId
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: since
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response2"
      summary: List
      tags:
      - Event Controller
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response5"
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Event Controller
  /events/api/events/names:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response1"
      summary: Names
      tags:
      - Event Controller
  /events/api/events/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishEventRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          description: Bad Request
      summary: Publish
      tags:
      - Event Controller
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response3"
      summary: Get
      tags:
      - Event Controller
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response4"
      summary: Delete
      tags:
      - Event Controller
info:
  title: qits-events API
  version: 2026.807.123727